-BIG UPDATE!!! - VERSION 0.05 ALPHA

+++ src/main/java/com/example/signtest/DataSet.java
@@ -3,17 +3,17 @@ package com.example.signtest;
 import java.util.ArrayList;
 
 public class DataSet {
-    private static ArrayList<Integer> x;
-    private static ArrayList<Integer> y;
+    private static ArrayList<Double> x;
+    private static ArrayList<Double> y;
 
-    public DataSet (ArrayList<Integer> data1, ArrayList<Integer> data2) {
+    public DataSet (ArrayList<Double> data1, ArrayList<Double> data2) {
         x = data1;
         y = data2;
     }
-    public static ArrayList<Integer> getData1() {
+    public static ArrayList<Double> getData1() {
         return x;
     }
-    public static ArrayList<Integer> getData2() {
+    public static ArrayList<Double> getData2() {
         return y;
     }
 }
